+++ index copy.html
@@ -27,17 +27,16 @@
                 <div id="restaurantTabMenuBox" class="restaurant-tab-menu-box">
                   <!-- 탭메뉴 렌더링 -->
                 </div>
+                <div id="restaurantFilterBox" class="restaurant-filter-box">
+                  <!-- 필터박스 렌더링 -->
+                </div>
+                <div id="restaurantListBox" class="restaurant-list-box">
+                  <!-- 런치리스트 렌더링 -->
+                </div>
               </section>
               <div class="restaurant-section-slider">
                 <div class="restaurant-section-container">
-                  <section id="restaurantListSection" class="restaurant-list-container">
-                    <div id="restaurantFilterBox" class="restaurant-filter-box">
-                      <!-- 필터박스 렌더링 -->
-                    </div>
-                    <div id="restaurantListBox" class="restaurant-list-box">
-                      <!-- 런치리스트 렌더링 -->
-                    </div>
-                  </section>
+                  <section id="restaurantListSection" class="restaurant-list-container"></section>
                   <section id="restaurantFavoriteSection" class="restaurant-list-container"></section>
                 </div>
               </div>
