+++ src/main.js
@@ -1,13 +1,14 @@
-import { LunchList } from "./components/LunchList.ts";
-import { RestaurantTabMenu } from "./components/RestaurantTabMenu.js";
+import { FilterBox } from "./components/function/FilterBox.js";
+import { LunchList } from "./components/function/LunchList.ts";
+import { RestaurantTabMenu } from "./components/function/RestaurantTabMenu.js";
 import SubmitEvent from "./utils/SubmitEvent.js";
 import { getHTML } from "./utils/utils.ts";
 
-const restaurantTabMenu = RestaurantTabMenu("restaurantMenuSection");
-const lunchList = LunchList(
-  "restaurantListSection",
-  "restaurantFavoriteSection"
-);
+const restaurantTabMenu = RestaurantTabMenu("restaurantTabMenuBox");
+const filterBox = FilterBox();
+console.log(filterBox);
+document.getElementById("restaurantFilterBox").innerHTML = filterBox;
+const lunchList = LunchList("restaurantListBox", "restaurantFavoriteSection");
 
 lunchList.render();
 lunchList.renderFavorites();
